requirements.txt
google-generativeai
langchain
scikit-learn
pandas
numpy
torch
tqdm
python-dotenv 